requirements.txt
# Core AI and ML libraries
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0

# Financial data APIs
yfinance>=0.2.20
alpha-vantage>=2.3.1
requests>=2.31.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web scraping and RSS feeds
beautifulsoup4>=4.12.0
feedparser>=6.0.10

# Natural language processing
nltk>=3.8.1
textblob>=0.17.1

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0